+++ src/test/java/com/ojeda/obras/web/rest/DetalleListaPrecioResourceIT.java
@@ -49,9 +49,9 @@ class DetalleListaPrecioResourceIT {
     private static final String DEFAULT_PRODUCT = "AAAAAAAAAA";
     private static final String UPDATED_PRODUCT = "BBBBBBBBBB";
 
-    private static final Float DEFAULT_AMOUNT = 1F;
-    private static final Float UPDATED_AMOUNT = 2F;
-    private static final Float SMALLER_AMOUNT = 1F - 1F;
+    private static final Double DEFAULT_AMOUNT = 1D;
+    private static final Double UPDATED_AMOUNT = 2D;
+    private static final Double SMALLER_AMOUNT = 1D - 1D;
 
     private static final String ENTITY_API_URL = "/api/detalle-lista-precios";
     private static final String ENTITY_API_URL_ID = ENTITY_API_URL + "/{id}";
